+++ src/test/CompleteSignUp.test.tsx
@@ -1,8 +1,9 @@
-import { render, screen } from '@testing-library/react';
+/* eslint-disable no-undef */
+/* eslint-disable import/no-extraneous-dependencies */
+import { render, screen, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { RecoilRoot } from 'recoil';
-import { MemoryRouter } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import THEME from '@/styles/theme';
 
@@ -31,37 +32,46 @@ describe('일반 가입 컴포넌트 테스트', () => {
     render(
       <RecoilRoot>
         <ThemeProvider theme={THEME}>
-          <MemoryRouter initialEntries={['/signup']}>
-            <Modal>
-              <CompleteSignUp id={id} />
-            </Modal>
-          </MemoryRouter>
+          <Modal>
+            <CompleteSignUp id={id} />
+          </Modal>
         </ThemeProvider>
       </RecoilRoot>,
     );
-  };
-
-  test('컴포넌트 렌더링', async () => {
-    rendering();
     const heading = screen.getByText(/회원가입이 완료되었습니다!/i);
-    expect(heading).toBeInTheDocument();
-  });
-
-  test('메인으로 이동 버튼을 클릭하면 메인페이지로 이동', async () => {
-    rendering();
     const signUpButton = screen.getByRole('button', {
       name: /메인으로/i,
     }) as HTMLButtonElement;
 
+    return {
+      heading,
+      signUpButton,
+    };
+  };
+
+  test('컴포넌트 렌더링', async () => {
+    const { heading } = rendering();
+    expect(heading).toBeInTheDocument();
+  });
+
+  test('메인으로 이동 버튼을 클릭하면 메인페이지로 이동', async () => {
+    const { signUpButton } = rendering();
+
     await userEvent.click(signUpButton);
+
     expect(mockedNavigate).toBeCalledTimes(1);
+    mockedNavigate.mockClear();
   });
 
   test('3초가 지나면 자동으로 메인페이지로 이동', async () => {
+    jest.useFakeTimers();
     rendering();
 
-    setTimeout(() => {
-      expect(mockedNavigate).toBeCalledTimes(1);
-    }, 3000);
+    act(() => {
+      jest.advanceTimersByTime(3000);
+    });
+    expect(mockedNavigate).toBeCalledTimes(1);
+
+    jest.useRealTimers();
   });
 });
